requirements.txt
matplotlib==3.3.2
numpy==1.21.0
pandas==1.1.4
pandas-datareader==0.9.0
pathlib==1.0.1
pickleshare==0.7.5
python-telegram-bot==13.8.1
ta==0.7.0